# JsonGenerator/endpoint-server/openapi/swaggerInit.yaml
...
openapi: 3.1.0
info:
  title: GeoRDFBench Framework JSON Specification API with Swagger
  version: 2.0.0
  description: A cRud API application made with Express and documented with Swagger
  license:
    name: Apache-2.0
    url: https://spdx.org/licenses/Apache-2.0.html
  contact:
    name: Theofilos Ioannidis
    url: http://users.uoa.gr/~tioannid/
    email: tioannid@di.uoa.gr
servers:
  - url: http://localhost:{port}/
    description: Development server
    variables:
      port:
        enum:
          - "5000"
          - "3000"
        default: "5000"
  - url: "{scheme}://geordfbench.di.uoa.gr/jsonapi/"
    description: Production server
    variables:
      scheme:
        enum:
          - http
          - https
        default: http
paths:
...
  "/datasets/{spec}":
    get:
      summary: Get the dataset by specification name
      tags:
        - Datasets
      parameters:
        - in: path
          name: spec
          schema:
            type: string
          required: true
          description: The dataset specification name
      responses:
        "200":
          description: The specification for the dataset specification name provided
          contens:
            application/json:
              schema:
                "$ref": "#/components/schemas/Dataset"
        "404":
          description: The dataset specification name was not found
  "/datasets/{newspec}":
    post:
      summary: Create the dataset by specification name
      requestBody:
        description: Specify a **JSON** string
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/Dataset"
      tags:
        - Datasets
      parameters:
        - in: path
          name: newspec
          schema:
            type: string
          required: true
          description: The dataset specification name
      responses:
        "201":
          description: The dataset specification was created
          contens:
            application/json:
              schema:
                "$ref": "#/components/schemas/Dataset"
        "400":
          description: A field is either missing or the n field is not integer!
        "404":
          description: The dataset specification name already exists!
  "/datasets/{existingspec}":
    delete:
      summary: Delete an existing host by specification name
      tags:
        - Datasets
      parameters:
        - in: path
          name: existingspec
          schema:
            type: string
          required: true
          description: The host specification name
      responses:
        "204":
          description: The host specification was deleted
        "404":
          description: The host specification name was not found
...
  "/executionspecs/{spec}":
    get:
      summary: Get the execution by specification name
      tags:
        - Execution Specifications
      parameters:
        - in: path
          name: spec
          schema:
            type: string
          required: true
          description: The execution specification name
      responses:
        "200":
          description: The specification for the execution specification name provided
          contens:
            application/json:
              schema:
                "$ref": "#/components/schemas/ExecutionSpec"
        "404":
          description: The dataset specification name was not found
  "/executionspecs/{newspec}":
    post:
      summary: Create the execution specification by specification name
      requestBody:
        description: Specify a **JSON** string
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/ExecutionSpec"
      tags:
        - Execution Specifications
      parameters:
        - in: path
          name: newspec
          schema:
            type: string
          required: true
          description: The execution specification name
      responses:
        "201":
          description: The execution specification was created
          contens:
            application/json:
              schema:
                "$ref": "#/components/schemas/ExecutionSpec"
        "400":
          description: A field is missing!
        "403":
          description:
            "Not integer value one of: { maxDurationSecsPerQueryRep | maxDurationSecs
            | clearCacheDelaymSecs }"
        "404":
          description: The execution specification name already exists!
  "/executionspecs/{existingspec}":
    delete:
      summary: Delete an existing execution specification by specification name
      tags:
        - Execution Specifications
      parameters:
        - in: path
          name: existingspec
          schema:
            type: string
          required: true
          description: The execution specification name
      responses:
        "204":
          description: The execution specification was deleted
        "404":
          description: The execution specification name was not found
...
components:
  schemas:
    Dataset:
      type: object
      required:
        - classname
        - name
        - relativeBaseDir
        - simpleGeospatialDataSetList
        - mapDataSetContexts
        - "n"
      properties:
        classname:
          type: string
          description: The java full class name
        name:
          type: string
          description: The name of the dataset
        relativeBaseDir:
          type: string
          format: path
          description: The actual dataset's path relative to the host's sourceFileDir
        simpleGeospatialDataSetList:
          type: object
          description: A single file dataset
        mapDataSetContexts:
          type: array
          description: The named graphs (contexts) used in the simple dataset files
        "n":
          type: number
          format: int32
          description: Scaling factor for synthetially generated datasets
      example:
        classname: gr.uoa.di.rdf.geordfbench.runtime.datasets.complex.impl.GeographicaDS
        name: scalability_500M
        relativeBaseDir: Scalability/500M
        simpleGeospatialDataSetList:
          name: scalability_500M
          relativeBaseDir: Scalability/500M
          dataFile: scalability_500M.nt
          rdfFormat: N-TRIPLES
          mapUsefulNamespacePrefixes:
            geo: "<http://www.opengis.net/ont/geosparql#>"
            rdf: "<http://www.w3.org/1999/02/22-rdf-syntax-ns#>"
            owl: "<http://www.w3.org/2002/07/owl#>"
            geof: "<http://www.opengis.net/def/function/geosparql/>"
            lgo: "<http://data.linkedeodata.eu/ontology#>"
            xsd: "<http://www.w3.org/2001/XMLSchema#>"
            rdfs: "<http://www.w3.org/2000/01/rdf-schema#>"
            geo-sf: "<http://www.opengis.net/ont/sf#>"
          mapAsWKT:
            scalabilityAsWKT: "<http://www.opengis.net/ont/geosparql#asWKT>"
          mapHasGeometry:
            scalabilityHasGeometry: "<http://www.opengis.net/ont/geosparql#hasGeometry>"
        mapDataSetContexts:
          scalability_500M:
        "n": 0
...
    ExecutionSpec:
      type: object
      required:
        - classname
        - execTypeReps
        - maxDurationSecsPerQueryRep
        - maxDurationSecs
        - action
        - avgFunc
        - onColdFailure
        - clearCacheDelaymSecs
      properties:
        classname:
          type: string
          description: The java full class name
        execTypeReps:
          type: array
          description: The repetitions per execution type
        maxDurationSecsPerQueryRep:
          type: number
          format: int32
          description: The max duration (secs) per query execution
        maxDurationSecs:
          type: number
          format: int32
          description: The max duration (secs) for the queryset execution
        action:
          type: string
          enum:
            - RUN
            - PRINT
          description: The action that experiment should perform
        avgFunc:
          type: string
          enum:
            - QUERY_MEDIAN
            - QUERY_MEAN
            - QUERYSET_MEAN
          description: The average function and the scope of its use
        onColdFailure:
          type: string
          enum:
            - SKIP_REMAINING_ALL_QUERY_EXECUTIONS
            - SKIP_REMAINING_COLD_EXECUTIONS
            - EXCLUDE_AND_PROCEED
          description: Desired behaviour upon COLD execution failure
        clearCacheDelaymSecs:
          type: number
          format: int32
          description: The delay (msecs) waiting for clearing caches and garbage collection
      example:
        classname: gr.uoa.di.rdf.geordfbench.runtime.executionspecs.impl.SimpleES
        execTypeReps:
          COLD: 3
          WARM: 3
        maxDurationSecsPerQueryRep: 86400
        maxDurationSecs: 604800
        action: RUN
        avgFunc: QUERY_MEDIAN
        onColdFailure: SKIP_REMAINING_ALL_QUERY_EXECUTIONS
        clearCacheDelaymSecs: 5000
...